requirements.txt
virtualenv
django
djangorestframework
Pillow
djangorestframework-simplejwt
django-cors-headers
django-environ